[iOS] fix: TrackInfoView를 새 파일로 분리
