+++ iOS/MiniVibe/MiniVibe/Scenes/TrackList/TrackCellView.swift
@@ -59,28 +59,6 @@ struct Ellipsis: View {
     }
 }
 
-struct TrackInfoView: View {
-    let title: String
-    let artist: String
-    let coverURLString: String?
-    
-    var body: some View {
-        HStack {
-            AsyncImage(url: URL(string: coverURLString ?? ""))
-                .frame(width: 44, height: 44, alignment: .center)
-                .padding(.vertical, 2)
-            VStack(alignment: .leading) {
-                Text(title)
-                    .modifier(Title2())
-                Text(artist)
-                    .modifier(Description2())
-            }
-            Spacer()
-        }
-    }
-}
-
-
 struct TrackCellView_Previews: PreviewProvider {
     
     static var previews: some View {
